content-deletion - fixes for handling deletion of repos... and deletion of content from last env

+++ app/models/deletion_changeset.rb
@@ -114,7 +114,13 @@ class DeletionChangeset < Changeset
       product = repo.product
       next if (products.uniq! or []).include? product
 
-      product.delete_repo_by_id(repo.id)
+      # if the repo is being deleted from the last environment in the path, delete it; otherwise,
+      # mark it as 'disabled'... this will hide the repo...
+      if (from_env.successor.nil?)
+        product.delete_repo_by_id(repo.id)
+      else
+        repo.disable_repo
+      end
     end
   end
